requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
openai>=1.20.0
pyyaml>=6.0.1
sentence-transformers>=2.6.0
numpy>=1.26.0
httpx>=0.27.0
pytest>=8.1.0
pytest-asyncio>=0.23.0
